+++ src/controller/book/book.controller.ts
@@ -1,12 +1,15 @@
 import { Request, Response } from "express";
 import * as bookRepository from "../../repositories/book/book.repository";
 import { bookTypeEnum } from "../../db/schema";
+import logger from "../../utils/logger";
 
 export const getBookHandler = async (req: Request, res: Response) => {
   try {
     const bookType = req.params.bookType as string;
 
     if (!bookType) {
+      logger.warn("Pencarian gagal: Parameter tipe buku kosong");
+
       return res.status(400).json({
         success: false,
         message: "Buku tidak ditemukan",
@@ -14,6 +17,8 @@ export const getBookHandler = async (req: Request, res: Response) => {
     }
 
     if (!bookTypeEnum.enumValues.includes(bookType as any)) {
+      logger.warn(`Pencarian gagal: Tipe buku tidak valid - ${bookType}`);
+
       return res.status(400).json({
         success: false,
         message: "Tipe buku tidak valid",
@@ -24,7 +29,16 @@ export const getBookHandler = async (req: Request, res: Response) => {
     const limit = Math.max(1, parseInt(req.query.limit as string) || 10);
     const search = (req.query.search as string) || "";
 
-    const result = await bookRepository.findBooksWithPagination(bookType, page, limit, search);
+    const result = await bookRepository.findBooksWithPagination(
+      bookType,
+      page,
+      limit,
+      search,
+    );
+
+    logger.info(
+      `Memproses request data buku: tipe=${bookType}, search=${search}, page=${page}`,
+    );
 
     return res.status(200).json({
       success: true,
@@ -32,6 +46,11 @@ export const getBookHandler = async (req: Request, res: Response) => {
       ...result,
     });
   } catch (error) {
+    logger.error(
+      `Error pada getBookHandler: ${error instanceof Error ? error.message : "Unknown Error"}`,
+      { error },
+    );
+
     return res.status(500).json({
       success: false,
       message: "Internal server error",