+++ src/controller/shelf/stack.controller.ts
@@ -1,11 +1,14 @@
 import { Request, Response } from "express";
 import * as stackRepository from "../../repositories/shelf/stack.repository";
+import logger from "../../utils/logger";
 
 export const getStackHandler = async (req: Request, res: Response) => {
   try {
     const shelfId = req.params.id as string;
 
     if (!shelfId) {
+      logger.warn("Pencarian gagal: Parameter rak id kosong");
+
       return res.status(404).json({
         success: false,
         message: "Data not found",
@@ -20,7 +23,11 @@ export const getStackHandler = async (req: Request, res: Response) => {
       shelfId,
       page,
       limit,
-      search
+      search,
+    );
+
+    logger.info(
+      `Memproses request data susunan rak: rak_id=${shelfId}, search=${search}, page=${page}`,
     );
 
     return res.status(200).json({
@@ -29,6 +36,11 @@ export const getStackHandler = async (req: Request, res: Response) => {
       ...result,
     });
   } catch (error) {
+    logger.error(
+      `Error pada getStackHandler: ${error instanceof Error ? error.message : "Unknown Error"}`,
+      { error },
+    );
+
     return res.status(500).json({
       success: false,
       message: "Internal server error",