+++ src/controller/transaction/transaction.controller.ts
@@ -1,12 +1,15 @@
 import { Request, Response } from "express";
 import * as transactionRepository from "../../repositories/transaction/transaction.repository";
 import { transactionStatusEnum } from "../../db/schema";
+import logger from "../../utils/logger";
 
 export const getTransactionHandler = async (req: Request, res: Response) => {
   try {
     const status = req.params.status as string;
 
     if (!status) {
+      logger.warn("Pencarian gagal: Parameter status kosong");
+
       return res.status(400).json({
         success: false,
         message: "Status tidak ditemukan",
@@ -17,6 +20,8 @@ export const getTransactionHandler = async (req: Request, res: Response) => {
       status !== "Semua" &&
       !transactionStatusEnum.enumValues.includes(status as any)
     ) {
+      logger.warn(`Pencarian gagal: Tipe transaksi tidak valid - ${status}`);
+
       return res.status(400).json({
         success: false,
         message: "Status transaksi tidak valid",
@@ -31,7 +36,11 @@ export const getTransactionHandler = async (req: Request, res: Response) => {
       status,
       page,
       limit,
-      search
+      search,
+    );
+
+    logger.info(
+      `Memproses request data transaksi: status=${status}, search=${search}, page=${page}`,
     );
 
     return res.status(200).json({
@@ -40,6 +49,11 @@ export const getTransactionHandler = async (req: Request, res: Response) => {
       ...result,
     });
   } catch (error) {
+    logger.error(
+      `Error pada getTransactionHandler: ${error instanceof Error ? error.message : "Unknown Error"}`,
+      { error },
+    );
+
     return res.status(500).json({
       success: false,
       message: "Internal server error",