+++ src/repositories/shelf/shelf.repository.ts
@@ -2,6 +2,7 @@ import { count, ilike, desc } from "drizzle-orm";
 import { db } from "../../db";
 import { shelves } from "../../db/schema";
 import redisClient from "../../config/redis";
+import logger from "../../utils/logger";
 
 export async function findShelvesWithPagination(
   page: number = 1,
@@ -12,8 +13,12 @@ export async function findShelvesWithPagination(
   const cachedData = await redisClient.get(cacheKey);
 
   if (cachedData) {
+    logger.info(`Cache hit: Mengambil data dari Redis untuk key ${cacheKey}`);
+
     return JSON.parse(cachedData);
   }
+
+  logger.info(`Cache miss: Mengambil data dari Database untuk key ${cacheKey}`);
 
   const offset = (page - 1) * limit;
   const whereClause = search
@@ -47,6 +52,8 @@ export async function findShelvesWithPagination(
   };
 
   await redisClient.setEx(cacheKey, 60, JSON.stringify(result));
+
+  logger.info(`Data baru berhasil disimpan ke Redis untuk key ${cacheKey}`);
 
   return result;
 }