+++ src/repositories/user/member.repository.ts
@@ -2,6 +2,7 @@ import { count, or, ilike, eq, and, desc } from "drizzle-orm";
 import { db } from "../../db";
 import { members } from "../../db/schema";
 import redisClient from "../../config/redis";
+import logger from "../../utils/logger";
 
 export async function findMembersWithPagination(
   statusActive: string,
@@ -13,8 +14,12 @@ export async function findMembersWithPagination(
   const cachedData = await redisClient.get(cacheKey);
 
   if (cachedData) {
+    logger.info(`Cache hit: Mengambil data dari Redis untuk key ${cacheKey}`);
+
     return JSON.parse(cachedData);
   }
+
+  logger.info(`Cache miss: Mengambil data dari Database untuk key ${cacheKey}`);
 
   const offset = (page - 1) * limit;
 
@@ -61,6 +66,8 @@ export async function findMembersWithPagination(
   };
 
   await redisClient.setEx(cacheKey, 60, JSON.stringify(result));
+
+  logger.info(`Data baru berhasil disimpan ke Redis untuk key ${cacheKey}`);
 
   return result;
 }